Move format_journey_node_guideline_id from engine common to projection
