+++ src/parlant/core/engines/alpha/guideline_matching/generic/common.py
@@ -16,8 +16,7 @@ from dataclasses import dataclass
 import json
 from typing import Optional, cast
 
-from parlant.core.guidelines import Guideline, GuidelineId
-from parlant.core.journeys import JourneyEdgeId, JourneyLinkId, JourneyNodeId
+from parlant.core.guidelines import Guideline
 
 
 @dataclass
@@ -46,17 +45,3 @@ def internal_representation(g: Guideline) -> GuidelineInternalRepresentation:
         action = cast(str, internal_action) or action
 
     return GuidelineInternalRepresentation(condition, action, description)
-
-
-def format_journey_node_guideline_id(
-    node_id: JourneyNodeId,
-    edge_id: Optional[JourneyEdgeId] = None,
-    link_id: Optional[JourneyLinkId] = None,
-) -> GuidelineId:
-    if edge_id and link_id:
-        return GuidelineId(f"journey_node:{node_id}:{edge_id}:{link_id}")
-
-    if edge_id:
-        return GuidelineId(f"journey_node:{node_id}:{edge_id}")
-
-    return GuidelineId(f"journey_node:{node_id}")
