+++ src/parlant/sdk.py
@@ -112,9 +112,7 @@ from parlant.core.context_variables import (
     ContextVariableStore,
 )
 from parlant.core.emission.event_publisher import EventPublisherFactory
-from parlant.core.engines.alpha.guideline_matching.generic.common import (
-    format_journey_node_guideline_id,
-)
+from parlant.core.journey_guideline_projection import format_journey_node_guideline_id
 from parlant.core.meter import Meter
 from parlant.core.tracer import Tracer
 from parlant.core.customers import (
@@ -223,7 +221,7 @@ from parlant.core.journeys import (
     JourneyNodeId,
     JourneyStore,
     JourneyVectorStore,
-    NodeKind,
+    JourneyNodeKind,
 )
 
 from parlant.core.loggers import LogLevel, Logger
@@ -2292,9 +2290,9 @@ class Journey:
         labels: Iterable[str] = (),
     ) -> TState:
         node_kind = {
-            ForkJourneyState: NodeKind.FORK,
-            ToolJourneyState: NodeKind.TOOL,
-            ChatJourneyState: NodeKind.CHAT,
+            ForkJourneyState: JourneyNodeKind.FORK,
+            ToolJourneyState: JourneyNodeKind.TOOL,
+            ChatJourneyState: JourneyNodeKind.CHAT,
         }[state_type]
 
         await _enable_tool_refs(self._server._plugin_server, tools)
